ofxIldaPolyProcessor: changed for loop counter to size_t instead of int to avoid warnings.

+++ src/ofxIldaPolyProcessor.h
@@ -49,7 +49,7 @@ namespace ofxIlda {
             float totalLength = 0;
             vector<int> pathLengths;
             processedPolys = origPolys;
-            for(int i=0; i<processedPolys.size(); i++) {
+            for(size_t i=0; i<processedPolys.size(); i++) {
                 if(processedPolys[i].size()) {
                     // smooth paths
                     if(params.smoothAmount > 0) processedPolys[i].setFromPolyline(processedPolys[i].getSmoothed(params.smoothAmount));
@@ -77,7 +77,7 @@ namespace ofxIlda {
             
             // resample paths based on spacing (either as calculated by targetPointCount, or set by user)
             if(params.spacing) {
-                for(int i=0; i<processedPolys.size(); i++) {
+                for(size_t i=0; i<processedPolys.size(); i++) {
                     processedPolys[i].setFromPolyline(processedPolys[i].getResampledBySpacing(params.spacing));
                 }
             }
